Move pcap write code out of UI.
