+++ src/main.rs
@@ -20,8 +20,8 @@ mod decoder;
 mod expander;
 mod id;
 mod index_stream;
-mod loader;
 mod model;
+mod pcap;
 mod rcu;
 mod row_data;
 mod stream;
