+++ src/pcap.rs
@@ -1,14 +1,16 @@
+use std::borrow::Cow;
 use std::fs::File;
-use std::io::BufReader;
+use std::io::{BufReader, BufWriter, Write};
 use std::mem::size_of;
 use std::path::PathBuf;
 
 use pcap_file::{
-    pcap::{PcapReader, PcapHeader, RawPcapPacket},
+    pcap::{PcapReader, PcapWriter, PcapHeader, RawPcapPacket},
+    DataLink,
     TsResolution,
 };
 
-use anyhow::Error;
+use anyhow::{Context, Error};
 
 pub struct Loader {
     pcap: PcapReader<BufReader<File>>,
@@ -16,6 +18,10 @@ pub struct Loader {
     pub bytes_read: u64,
     frac_ns: u64,
     start_time: Option<u64>,
+}
+
+pub struct Writer {
+    pcap: PcapWriter<BufWriter<File>>,
 }
 
 impl Loader {
@@ -53,5 +59,39 @@ impl Loader {
                 Some(Ok((packet, timestamp)))
             }
         }
+    }
+}
+
+impl Writer {
+    pub fn open(path: PathBuf) -> Result<Writer, Error> {
+        let file = File::create(path)?;
+        let writer = BufWriter::new(file);
+        let header = PcapHeader {
+            datalink: DataLink::USB_2_0,
+            ts_resolution: TsResolution::NanoSecond,
+            .. PcapHeader::default()
+        };
+        Ok(Writer{pcap: PcapWriter::with_header(writer, header)?})
+    }
+
+    pub fn add_packet(&mut self, bytes: &[u8], timestamp_ns: u64) -> Result<(), Error> {
+        let length: u32 = bytes
+            .len()
+            .try_into()
+            .context("Packet too large for pcap file")?;
+        let packet = RawPcapPacket {
+            ts_sec: (timestamp_ns / 1_000_000_000) as u32,
+            ts_frac: (timestamp_ns % 1_000_000_000) as u32,
+            incl_len: length,
+            orig_len: length,
+            data: Cow::from(bytes)
+        };
+        self.pcap.write_raw_packet(&packet)?;
+        Ok(())
+    }
+
+    pub fn close(self) -> Result<(), Error> {
+        self.pcap.into_writer().flush()?;
+        Ok(())
     }
 }
